# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "An OGC compliant process server.",
        "title": "Process-API Server",
        "contact": {
            "name": "Seth Lawler",
            "email": "slawler@dewberry.com"
        },
        "version": "dev-8.16.23"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/jobs/{jobID}": {
            "get": {
                "description": "[Job Status Specification](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_status_info)",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Job Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example: 44d9ca0e-2ca7-4013-907f-a8ccc60da3b4",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.jobResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "[Dismss Job Specification](https://docs.ogc.org/is/18-062r2/18-062r2.html#ats_dismiss)",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Dismiss Job",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.jobResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{jobID}/logs": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Job Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example: 44d9ca0e-2ca7-4013-907f-a8ccc60da3b4",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jobs.JobLogs"
                        }
                    }
                }
            }
        },
        "/jobs/{jobID}/results": {
            "get": {
                "description": "Provides metadata associated with a job",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Job Metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example: 44d9ca0e-2ca7-4013-907f-a8ccc60da3b4",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/processes/{processID}": {
            "get": {
                "description": "[Process Description Specification](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description)",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "Describe Process Information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "example: pyecho",
                        "name": "processID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/processes.processDescription"
                        }
                    }
                }
            }
        },
        "/processes/{processID}/execution": {
            "post": {
                "description": "[Execute Process Specification](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_create_job)",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processes"
                ],
                "summary": "Execute Process",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pyecho",
                        "name": "processID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "example: {“text”:“Hello World!”}",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.jobResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.jobResponse": {
            "type": "object",
            "properties": {
                "jobID": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "outputs": {},
                "processID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "default": "process"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "jobs.JobLogs": {
            "type": "object",
            "properties": {
                "api_logs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "container_logs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "jobID": {
                    "type": "string"
                },
                "processID": {
                    "type": "string"
                }
            }
        },
...
        "processes.Info": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobControlOptions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "outputTransmission": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "processes.Input": {
            "type": "object",
            "properties": {
                "literalDataDomain": {
                    "$ref": "#/definitions/processes.LiteralDataDomain"
                }
            }
        },
        "processes.Inputs": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input": {
                    "$ref": "#/definitions/processes.Input"
                },
                "maxOccurs": {
                    "type": "integer"
                },
                "minOccurs": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "processes.Link": {
            "type": "object",
            "properties": {
                "href": {
                    "type": "string"
                },
                "rel": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "processes.LiteralDataDomain": {
            "type": "object",
            "properties": {
                "dataType": {
                    "type": "string"
                },
                "valueDefinition": {
                    "$ref": "#/definitions/processes.ValueDefinition"
                }
            }
        },
        "processes.Output": {
            "type": "object",
            "properties": {
                "formats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "processes.Outputs": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inputID": {
                    "description": "json omit",
                    "type": "string"
                },
                "output": {
                    "$ref": "#/definitions/processes.Output"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "processes.Resources": {
            "type": "object",
            "properties": {
                "cpus": {
                    "type": "number"
                },
                "memory": {
                    "type": "integer"
                }
            }
        },
...
        "processes.processDescription": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "info": {
                    "$ref": "#/definitions/processes.Info"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/processes.Inputs"
                    }
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/processes.Link"
                    }
                },
                "maxResources": {
                    "$ref": "#/definitions/processes.Resources"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/processes.Outputs"
                    }
                }
            }
        }
    }
}